make a custom hook useShareGroup to handle logic to share group
